test: 기존 EquipmentServiceImplTest에서 EquipmentItem 관련 책임을 분리
