+++ src/test/java/com/equip/equiprental/equipment/service/EquipmentServiceImplTest.java
@@ -6,14 +6,12 @@ import com.equip.equiprental.common.exception.CustomException;
 import com.equip.equiprental.common.exception.ErrorType;
 import com.equip.equiprental.equipment.domain.*;
 import com.equip.equiprental.equipment.dto.*;
-import com.equip.equiprental.equipment.repository.EquipmentItemHistoryRepository;
 import com.equip.equiprental.equipment.repository.EquipmentItemRepository;
 import com.equip.equiprental.equipment.repository.EquipmentRepository;
 import com.equip.equiprental.equipment.util.ModelCodeGenerator;
 import com.equip.equiprental.filestorage.domain.FileMeta;
 import com.equip.equiprental.filestorage.repository.FileRepository;
 import com.equip.equiprental.filestorage.service.FileService;
-import com.equip.equiprental.member.domain.Member;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.api.Test;
@@ -41,7 +39,6 @@ import static org.mockito.Mockito.*;
 public class EquipmentServiceImplTest {
     @Mock private EquipmentRepository equipmentRepository;
     @Mock private EquipmentItemRepository equipmentItemRepository;
-    @Mock private EquipmentItemHistoryRepository equipmentItemHistoryRepository;
     @Mock private ModelCodeGenerator modelCodeGenerator;
     @Mock private FileRepository fileRepository;
     @Mock private FileService fileService;
@@ -420,153 +417,6 @@ public class EquipmentServiceImplTest {
                     .isInstanceOf(CustomException.class)
                     .extracting("errorType")
                     .isEqualTo(ErrorType.EQUIPMENT_NOT_FOUND);
-        }
-    }
-
-    @Nested
-    @DisplayName("updateItemStatus 메서드 테스트")
-    class updateItemStatus {
-        @Test
-        @DisplayName("성공 - 장비 아이템 상태 변경 및 히스토리 저장")
-        void updateItemStatus_success() {
-            // given
-            Member changer = Member.builder()
-                    .memberId(1L)
-                    .name("Admin")
-                    .build();
-            UpdateItemStatusDto dto = UpdateItemStatusDto.builder()
-                    .equipmentItemId(1L)
-                    .newStatus("RENTED")
-                    .build();
-
-            EquipmentItem item = EquipmentItem.builder()
-                    .equipmentItemId(1L)
-                    .status(EquipmentStatus.AVAILABLE)
-                    .build();
-
-            when(equipmentItemRepository.findById(dto.getEquipmentItemId())).thenReturn(Optional.of(item));
-
-            // when
-            equipmentService.updateItemStatus(dto, changer);
-
-            // then
-            // 상태 변경 검증
-            assertThat(item.getStatus()).isEqualTo(EquipmentStatus.RENTED);
-
-            // 히스토리 저장 검증
-            ArgumentCaptor<EquipmentItemHistory> captor = ArgumentCaptor.forClass(EquipmentItemHistory.class);
-            verify(equipmentItemHistoryRepository).save(captor.capture());
-
-            EquipmentItemHistory savedHistory = captor.getValue();
-            assertThat(savedHistory.getItem()).isEqualTo(item);
-            assertThat(savedHistory.getOldStatus()).isEqualTo(EquipmentStatus.AVAILABLE);
-            assertThat(savedHistory.getNewStatus()).isEqualTo(EquipmentStatus.RENTED);
-            assertThat(savedHistory.getChangedBy()).isEqualTo(changer);
-        }
-
-        @Test
-        @DisplayName("예외 - 장비 아이템 존재하지 않음")
-        void updateItemStatus_itemNotFound() {
-            UpdateItemStatusDto dto = UpdateItemStatusDto.builder()
-                    .equipmentItemId(1L)
-                    .newStatus("RENTED")
-                    .build();
-
-            when(equipmentItemRepository.findById(dto.getEquipmentItemId())).thenReturn(Optional.empty());
-
-            assertThatThrownBy(() -> equipmentService.updateItemStatus(dto, null))
-                    .isInstanceOf(CustomException.class)
-                    .extracting("errorType")
-                    .isEqualTo(ErrorType.EQUIPMENT_ITEM_NOT_FOUND);
-        }
-    }
-
-    @Nested
-    @DisplayName("getItemHistory 메서드 테스트")
-    class getItemHistory {
-        @Test
-        @DisplayName("성공 - 장비 아이템 히스토리 조회")
-        void getItemHistory_success() {
-            // given
-            Long equipmentItemId = 1L;
-            SearchParamDto paramDto = SearchParamDto.builder()
-                    .page(1)
-                    .size(10)
-                    .build();
-            Pageable pageable = paramDto.getPageable();
-
-            EquipmentItemHistoryDto history1 = EquipmentItemHistoryDto.builder()
-                    .oldStatus("AVAILABLE")
-                    .newStatus("RENTED")
-                    .changedBy("Admin")
-                    .build();
-
-            EquipmentItemHistoryDto history2 = EquipmentItemHistoryDto.builder()
-                    .oldStatus("AVAILABLE")
-                    .newStatus("RENTED")
-                    .changedBy("Admin")
-                    .build();
-
-            // Page 생성
-            Page<EquipmentItemHistoryDto> mockPage = new PageImpl<>(List.of(history1, history2), pageable, 2);
-
-            // Mockito stub
-            when(equipmentItemHistoryRepository.findHistoriesByEquipmentItemId(equipmentItemId, pageable))
-                    .thenReturn(mockPage);
-
-            // when
-            PageResponseDto<EquipmentItemHistoryDto> result = equipmentService.getItemHistory(equipmentItemId, paramDto);
-
-            // then
-            assertThat(result.getContent()).hasSize(2)
-                    .extracting(EquipmentItemHistoryDto::getOldStatus, EquipmentItemHistoryDto::getNewStatus, EquipmentItemHistoryDto::getChangedBy)
-                    .containsExactly(
-                            tuple("AVAILABLE", "RENTED", "Admin"),
-                            tuple("AVAILABLE", "RENTED", "Admin")
-                    );
-
-            assertThat(result.getPage()).isEqualTo(pageable.getPageNumber() + 1);
-            assertThat(result.getSize()).isEqualTo(pageable.getPageSize());
-            assertThat(result.getTotalElements()).isEqualTo(2);
-            assertThat(result.getTotalPages()).isEqualTo(1);
-            assertThat(result.getNumberOfElements()).isEqualTo(2);
-
-            assertThat(result.isFirst()).isTrue();
-            assertThat(result.isLast()).isTrue();
-            assertThat(result.isEmpty()).isFalse();
-        }
-
-
-        @Test
-        @DisplayName("성공 - 히스토리 없음 (빈 페이지)")
-        void getItemHistory_empty() {
-            // given
-            Long equipmentItemId = 1L;
-            SearchParamDto paramDto = SearchParamDto.builder()
-                    .page(1)
-                    .size(10)
-                    .build();
-            Pageable pageable = paramDto.getPageable();
-
-            Page<EquipmentItemHistoryDto> emptyPage = Page.empty(pageable);
-            when(equipmentItemHistoryRepository.findHistoriesByEquipmentItemId(equipmentItemId, pageable))
-                    .thenReturn(emptyPage);
-
-            // when
-            PageResponseDto<EquipmentItemHistoryDto> result = equipmentService.getItemHistory(equipmentItemId, paramDto);
-
-            // then
-            assertThat(result.getContent()).isEmpty();
-            assertThat(result.isEmpty()).isTrue();
-
-            assertThat(result.getPage()).isEqualTo(pageable.getPageNumber() + 1);
-            assertThat(result.getSize()).isEqualTo(pageable.getPageSize());
-            assertThat(result.getTotalElements()).isEqualTo(0);
-            assertThat(result.getTotalPages()).isEqualTo(0);
-            assertThat(result.getNumberOfElements()).isEqualTo(0);
-
-            assertThat(result.isFirst()).isTrue();
-            assertThat(result.isLast()).isTrue();
         }
     }
